modify package name from httprequest to utils
