+++ internal/controller/request/observe.go
@@ -2,11 +2,13 @@ package request
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 
 	"github.com/arielsepton/provider-http/apis/request/v1alpha1"
 	"github.com/arielsepton/provider-http/internal/controller/request/requestgen"
 	"github.com/arielsepton/provider-http/internal/json"
+	"github.com/arielsepton/provider-http/internal/utils"
 	"github.com/crossplane/crossplane-runtime/pkg/logging"
 	"github.com/pkg/errors"
 )
@@ -14,6 +16,7 @@ import (
 const (
 	errObjectNotFound = "object wasn't created"
 	errNoGetMapping   = "forProvider doesn't contain GET mapping"
+	errStatusCode     = "received status code "
 )
 
 // isUpToDate checks whether desired spec up to date with the observed state for a given request
@@ -41,8 +44,11 @@ func (c *external) isUpToDate(ctx context.Context, cr *v1alpha1.Request) (bool, 
 		return false, errors.New(errObjectNotFound)
 	}
 
-	desiredState, err := desiredState(cr, c.logger)
+	if utils.IsHTTPError(res.StatusCode) {
+		return false, errors.New(fmt.Sprint(errStatusCode, res.StatusCode, " indicates an error. aborting"))
+	}
 
+	desiredState, err := desiredState(cr, c.logger)
 	if err != nil {
 		return false, err
 	}
@@ -50,7 +56,7 @@ func (c *external) isUpToDate(ctx context.Context, cr *v1alpha1.Request) (bool, 
 	// TODO (REL): check what happens if one of them is not a json.
 	responseBodyMap, _ := json.JsonStringToMap(res.ResponseBody)
 	desiredStateMap, _ := json.JsonStringToMap(desiredState)
-	return json.Contains(responseBodyMap, desiredStateMap) && isHTTPSuccess(res.StatusCode), nil
+	return json.Contains(responseBodyMap, desiredStateMap) && utils.IsHTTPSuccess(res.StatusCode), nil
 }
 
 func desiredState(cr *v1alpha1.Request, logger logging.Logger) (string, error) {
@@ -60,8 +66,4 @@ func desiredState(cr *v1alpha1.Request, logger logging.Logger) (string, error) {
 	}
 	requestDetails, err := requestgen.GenerateValidRequestDetails(*methodPutMapping, cr, logger)
 	return requestDetails.Body, err
-}
-
-func isHTTPSuccess(statusCode int) bool {
-	return statusCode >= 200 && statusCode < 300
 }
