+++ internal/controller/request/request.go
@@ -38,7 +38,7 @@ import (
 	apisv1alpha1 "github.com/arielsepton/provider-http/apis/v1alpha1"
 	httpClient "github.com/arielsepton/provider-http/internal/clients/http"
 	"github.com/arielsepton/provider-http/internal/controller/request/requestgen"
-	httpRequestsUtils "github.com/arielsepton/provider-http/internal/utils/http_request_utils"
+	"github.com/arielsepton/provider-http/internal/utils"
 )
 
 const (
@@ -113,7 +113,7 @@ func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.E
 		return nil, errors.Wrap(err, errProviderNotRetrieved)
 	}
 
-	h, err := c.newHttpClientFn(l, httpRequestsUtils.WaitTimeout(cr.Spec.ForProvider.WaitTimeout))
+	h, err := c.newHttpClientFn(l, utils.WaitTimeout(cr.Spec.ForProvider.WaitTimeout))
 	if err != nil {
 		return nil, errors.Wrap(err, errNewHttpClient)
 	}
@@ -157,7 +157,7 @@ func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 
 	return managed.ExternalObservation{
 		ResourceExists:    true,
-		ResourceUpToDate:  synced && !(httpRequestsUtils.ShouldRetry(cr.Spec.ForProvider.RollbackRetriesLimit, cr.Status.Failed) && !httpRequestsUtils.RetriesLimitReached(cr.Status.Failed, cr.Spec.ForProvider.RollbackRetriesLimit)),
+		ResourceUpToDate:  synced && !(utils.ShouldRetry(cr.Spec.ForProvider.RollbackRetriesLimit, cr.Status.Failed) && !utils.RetriesLimitReached(cr.Status.Failed, cr.Spec.ForProvider.RollbackRetriesLimit)),
 		ConnectionDetails: nil,
 	}, nil
 }
@@ -165,7 +165,7 @@ func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 func (c *external) deployAction(ctx context.Context, cr *v1alpha1.Request, method string, url string, body string, headers map[string][]string) error {
 	res, err := c.http.SendRequest(ctx, method, url, body, headers)
 
-	resource := &httpRequestsUtils.RequestResource{
+	resource := &utils.RequestResource{
 		Resource:       cr,
 		RequestContext: ctx,
 		HttpResponse:   res,
@@ -174,14 +174,14 @@ func (c *external) deployAction(ctx context.Context, cr *v1alpha1.Request, metho
 
 	if err != nil {
 		setErr := resource.SetError(err)
-		return httpRequestsUtils.SetRequestResourceStatus(*resource, setErr)
+		return utils.SetRequestResourceStatus(*resource, setErr)
 	}
 
 	setStatusCode := resource.SetStatusCode()
 	setHeaders := resource.SetHeaders()
 	setBody := resource.SetBody()
 
-	return httpRequestsUtils.SetRequestResourceStatus(*resource, setStatusCode, setHeaders, setBody)
+	return utils.SetRequestResourceStatus(*resource, setStatusCode, setHeaders, setBody)
 }
 
 func (c *external) Create(ctx context.Context, mg resource.Managed) (managed.ExternalCreation, error) {
