+++ internal/utils/set_status.go
@@ -1,4 +1,4 @@
-package httprequest
+package utils
 
 import (
 	"context"
